slight refactor of pallets
