+++ aleph-client/src/aleph_zero.rs
@@ -1,7 +1,7 @@
 #[allow(dead_code, unused_imports, non_camel_case_types)]
 pub mod api {
     use super::api as root_mod;
-    pub static PALLETS: [&str; 21usize] = [
+    pub static PALLETS: [&str; 22usize] = [
         "System",
         "RandomnessCollectiveFlip",
         "Scheduler",
@@ -23,6 +23,7 @@ pub mod api {
         "Contracts",
         "NominationPools",
         "Identity",
+        "SessionExt",
     ];
     #[derive(
         :: subxt :: ext :: codec :: Decode,
@@ -65,6 +66,8 @@ pub mod api {
         NominationPools(nomination_pools::Event),
         #[codec(index = 20)]
         Identity(identity::Event),
+        #[codec(index = 21)]
+        SessionExt(session_ext::Event),
     }
     pub mod system {
         use super::{root_mod, runtime_types};
@@ -735,10 +738,10 @@ pub mod api {
                         "Events",
                         vec![],
                         [
-                            100u8, 112u8, 231u8, 192u8, 184u8, 16u8, 59u8, 201u8, 45u8, 235u8,
-                            104u8, 44u8, 105u8, 65u8, 84u8, 78u8, 86u8, 161u8, 206u8, 209u8, 53u8,
-                            182u8, 73u8, 138u8, 118u8, 194u8, 88u8, 181u8, 76u8, 179u8, 207u8,
-                            180u8,
+                            33u8, 245u8, 241u8, 11u8, 6u8, 167u8, 244u8, 125u8, 236u8, 226u8, 75u8,
+                            108u8, 155u8, 129u8, 236u8, 188u8, 116u8, 96u8, 241u8, 15u8, 0u8,
+                            177u8, 166u8, 233u8, 139u8, 51u8, 142u8, 167u8, 118u8, 167u8, 47u8,
+                            27u8,
                         ],
                     )
                 }
@@ -1183,10 +1186,9 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            241u8, 199u8, 15u8, 208u8, 101u8, 207u8, 116u8, 231u8, 145u8, 101u8,
-                            182u8, 193u8, 52u8, 246u8, 230u8, 10u8, 238u8, 234u8, 131u8, 246u8,
-                            126u8, 123u8, 214u8, 187u8, 113u8, 138u8, 176u8, 161u8, 56u8, 208u8,
-                            119u8, 119u8,
+                            100u8, 40u8, 123u8, 51u8, 208u8, 56u8, 185u8, 99u8, 233u8, 204u8, 59u8,
+                            117u8, 214u8, 151u8, 26u8, 44u8, 37u8, 34u8, 131u8, 60u8, 82u8, 255u8,
+                            79u8, 3u8, 167u8, 171u8, 213u8, 150u8, 53u8, 180u8, 28u8, 146u8,
                         ],
                     )
                 }
@@ -1230,9 +1232,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            175u8, 81u8, 2u8, 204u8, 241u8, 89u8, 189u8, 119u8, 222u8, 128u8,
-                            246u8, 58u8, 92u8, 75u8, 87u8, 15u8, 102u8, 143u8, 109u8, 32u8, 4u8,
-                            22u8, 209u8, 144u8, 58u8, 173u8, 6u8, 45u8, 93u8, 242u8, 18u8, 170u8,
+                            13u8, 165u8, 31u8, 28u8, 100u8, 209u8, 122u8, 93u8, 66u8, 164u8, 66u8,
+                            240u8, 140u8, 38u8, 223u8, 208u8, 137u8, 87u8, 113u8, 101u8, 14u8,
+                            205u8, 215u8, 136u8, 68u8, 153u8, 186u8, 227u8, 101u8, 104u8, 87u8,
+                            194u8,
                         ],
                     )
                 }
@@ -1277,10 +1280,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            82u8, 60u8, 236u8, 221u8, 124u8, 178u8, 23u8, 207u8, 179u8, 59u8,
-                            206u8, 11u8, 132u8, 15u8, 153u8, 104u8, 203u8, 134u8, 201u8, 255u8,
-                            210u8, 193u8, 51u8, 196u8, 104u8, 134u8, 219u8, 117u8, 75u8, 120u8,
-                            23u8, 244u8,
+                            125u8, 132u8, 90u8, 45u8, 187u8, 114u8, 226u8, 2u8, 207u8, 194u8,
+                            240u8, 129u8, 100u8, 185u8, 183u8, 25u8, 206u8, 38u8, 35u8, 87u8,
+                            243u8, 231u8, 151u8, 169u8, 225u8, 203u8, 173u8, 141u8, 3u8, 49u8,
+                            239u8, 119u8,
                         ],
                     )
                 }
@@ -1311,10 +1314,9 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            145u8, 124u8, 34u8, 62u8, 145u8, 204u8, 132u8, 210u8, 101u8, 75u8,
-                            171u8, 66u8, 165u8, 150u8, 233u8, 214u8, 1u8, 84u8, 85u8, 140u8, 47u8,
-                            97u8, 45u8, 118u8, 235u8, 240u8, 38u8, 145u8, 132u8, 183u8, 123u8,
-                            157u8,
+                            243u8, 34u8, 254u8, 169u8, 4u8, 149u8, 74u8, 28u8, 5u8, 246u8, 62u8,
+                            114u8, 216u8, 6u8, 155u8, 229u8, 91u8, 204u8, 137u8, 156u8, 134u8, 6u8,
+                            193u8, 35u8, 118u8, 249u8, 151u8, 60u8, 13u8, 45u8, 186u8, 5u8,
                         ],
                     )
                 }
@@ -6286,10 +6288,10 @@ pub mod api {
                         "NextAuthorities",
                         vec![],
                         [
-                            120u8, 66u8, 194u8, 154u8, 194u8, 127u8, 75u8, 106u8, 59u8, 172u8,
-                            86u8, 242u8, 209u8, 110u8, 207u8, 132u8, 69u8, 73u8, 186u8, 235u8,
-                            95u8, 97u8, 117u8, 132u8, 76u8, 96u8, 86u8, 87u8, 74u8, 232u8, 91u8,
-                            120u8,
+                            39u8, 148u8, 4u8, 182u8, 70u8, 175u8, 165u8, 146u8, 170u8, 200u8,
+                            107u8, 146u8, 23u8, 217u8, 135u8, 183u8, 10u8, 207u8, 193u8, 178u8,
+                            199u8, 72u8, 43u8, 238u8, 70u8, 193u8, 126u8, 78u8, 155u8, 93u8, 158u8,
+                            208u8,
                         ],
                     )
                 }
@@ -6427,44 +6429,6 @@ pub mod api {
                 Eq,
                 PartialEq,
             )]
-            pub struct SetBanConfig {
-                pub minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
-                pub underperformed_session_count_threshold:
-                    ::core::option::Option<::core::primitive::u32>,
-                pub clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
-                pub ban_period: ::core::option::Option<::core::primitive::u32>,
-            }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
-            pub struct BanFromCommittee {
-                pub banned: ::subxt::ext::sp_core::crypto::AccountId32,
-                pub ban_reason: ::std::vec::Vec<::core::primitive::u8>,
-            }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
-            pub struct CancelBan {
-                pub banned: ::subxt::ext::sp_core::crypto::AccountId32,
-            }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
             pub struct SetElectionsOpenness {
                 pub openness: runtime_types::primitives::ElectionOpenness,
             }
@@ -6494,67 +6458,6 @@ pub mod api {
                             88u8, 2u8, 255u8, 219u8, 50u8, 103u8, 169u8, 150u8, 249u8, 161u8, 57u8,
                             39u8, 6u8, 241u8, 94u8, 139u8, 206u8, 236u8, 160u8, 92u8, 163u8, 170u8,
                             222u8, 99u8, 50u8, 91u8, 194u8, 192u8, 99u8, 123u8, 41u8, 136u8,
-                        ],
-                    )
-                }
-                #[doc = "Sets ban config, it has an immediate effect"]
-                pub fn set_ban_config(
-                    &self,
-                    minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
-                    underperformed_session_count_threshold: ::core::option::Option<
-                        ::core::primitive::u32,
-                    >,
-                    clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
-                    ban_period: ::core::option::Option<::core::primitive::u32>,
-                ) -> ::subxt::tx::StaticTxPayload<SetBanConfig> {
-                    ::subxt::tx::StaticTxPayload::new(
-                        "Elections",
-                        "set_ban_config",
-                        SetBanConfig {
-                            minimal_expected_performance,
-                            underperformed_session_count_threshold,
-                            clean_session_counter_delay,
-                            ban_period,
-                        },
-                        [
-                            228u8, 199u8, 170u8, 155u8, 208u8, 190u8, 211u8, 218u8, 105u8, 213u8,
-                            240u8, 152u8, 92u8, 19u8, 164u8, 28u8, 215u8, 145u8, 47u8, 248u8,
-                            219u8, 75u8, 234u8, 78u8, 29u8, 189u8, 35u8, 106u8, 165u8, 76u8, 27u8,
-                            50u8,
-                        ],
-                    )
-                }
-                #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
-                pub fn ban_from_committee(
-                    &self,
-                    banned: ::subxt::ext::sp_core::crypto::AccountId32,
-                    ban_reason: ::std::vec::Vec<::core::primitive::u8>,
-                ) -> ::subxt::tx::StaticTxPayload<BanFromCommittee> {
-                    ::subxt::tx::StaticTxPayload::new(
-                        "Elections",
-                        "ban_from_committee",
-                        BanFromCommittee { banned, ban_reason },
-                        [
-                            60u8, 254u8, 80u8, 201u8, 64u8, 189u8, 255u8, 111u8, 14u8, 9u8, 68u8,
-                            177u8, 196u8, 107u8, 10u8, 177u8, 78u8, 134u8, 98u8, 21u8, 179u8, 9u8,
-                            111u8, 185u8, 155u8, 39u8, 148u8, 88u8, 239u8, 16u8, 24u8, 171u8,
-                        ],
-                    )
-                }
-                #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
-                pub fn cancel_ban(
-                    &self,
-                    banned: ::subxt::ext::sp_core::crypto::AccountId32,
-                ) -> ::subxt::tx::StaticTxPayload<CancelBan> {
-                    ::subxt::tx::StaticTxPayload::new(
-                        "Elections",
-                        "cancel_ban",
-                        CancelBan { banned },
-                        [
-                            103u8, 192u8, 40u8, 246u8, 206u8, 52u8, 222u8, 51u8, 39u8, 247u8,
-                            220u8, 175u8, 232u8, 31u8, 168u8, 99u8, 206u8, 45u8, 191u8, 161u8,
-                            107u8, 12u8, 112u8, 54u8, 163u8, 170u8, 221u8, 220u8, 122u8, 177u8,
-                            178u8, 246u8,
                         ],
                     )
                 }
@@ -6598,39 +6501,6 @@ pub mod api {
             impl ::subxt::events::StaticEvent for ChangeValidators {
                 const PALLET: &'static str = "Elections";
                 const EVENT: &'static str = "ChangeValidators";
-            }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
-            #[doc = "Ban thresholds for the next era has changed"]
-            pub struct SetBanConfig(pub runtime_types::primitives::BanConfig);
-            impl ::subxt::events::StaticEvent for SetBanConfig {
-                const PALLET: &'static str = "Elections";
-                const EVENT: &'static str = "SetBanConfig";
-            }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
-            #[doc = "Validators have been banned from the committee"]
-            pub struct BanValidators(
-                pub  ::std::vec::Vec<(
-                    ::subxt::ext::sp_core::crypto::AccountId32,
-                    runtime_types::primitives::BanInfo,
-                )>,
-            );
-            impl ::subxt::events::StaticEvent for BanValidators {
-                const PALLET: &'static str = "Elections";
-                const EVENT: &'static str = "BanValidators";
             }
         }
         pub mod storage {
@@ -6747,186 +6617,6 @@ pub mod api {
                         ],
                     )
                 }
-                #[doc = " A lookup how many blocks a validator produced."]
-                pub fn session_validator_block_count(
-                    &self,
-                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "SessionValidatorBlockCount",
-                        vec![::subxt::storage::address::StorageMapKey::new(
-                            _0.borrow(),
-                            ::subxt::storage::address::StorageHasher::Twox64Concat,
-                        )],
-                        [
-                            186u8, 91u8, 185u8, 144u8, 216u8, 179u8, 157u8, 132u8, 17u8, 247u8,
-                            241u8, 172u8, 32u8, 7u8, 28u8, 60u8, 188u8, 192u8, 64u8, 29u8, 153u8,
-                            100u8, 130u8, 245u8, 189u8, 251u8, 68u8, 161u8, 202u8, 29u8, 153u8,
-                            131u8,
-                        ],
-                    )
-                }
-                #[doc = " A lookup how many blocks a validator produced."]
-                pub fn session_validator_block_count_root(
-                    &self,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                    (),
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "SessionValidatorBlockCount",
-                        Vec::new(),
-                        [
-                            186u8, 91u8, 185u8, 144u8, 216u8, 179u8, 157u8, 132u8, 17u8, 247u8,
-                            241u8, 172u8, 32u8, 7u8, 28u8, 60u8, 188u8, 192u8, 64u8, 29u8, 153u8,
-                            100u8, 130u8, 245u8, 189u8, 251u8, 68u8, 161u8, 202u8, 29u8, 153u8,
-                            131u8,
-                        ],
-                    )
-                }
-                #[doc = " Total possible reward per validator for the current era."]
-                pub fn validator_era_total_reward(
-                    &self,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<
-                        runtime_types::pallet_elections::ValidatorTotalRewards<
-                            ::subxt::ext::sp_core::crypto::AccountId32,
-                        >,
-                    >,
-                    ::subxt::storage::address::Yes,
-                    (),
-                    (),
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "ValidatorEraTotalReward",
-                        vec![],
-                        [
-                            111u8, 103u8, 48u8, 14u8, 23u8, 139u8, 162u8, 122u8, 212u8, 85u8, 64u8,
-                            188u8, 36u8, 142u8, 80u8, 224u8, 89u8, 63u8, 104u8, 86u8, 51u8, 111u8,
-                            166u8, 53u8, 189u8, 181u8, 240u8, 250u8, 160u8, 128u8, 179u8, 9u8,
-                        ],
-                    )
-                }
-                #[doc = " Current era config for ban functionality, see [`BanConfig`]"]
-                pub fn ban_config(
-                    &self,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanConfig>,
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                    (),
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "BanConfig",
-                        vec![],
-                        [
-                            115u8, 228u8, 135u8, 32u8, 85u8, 156u8, 44u8, 195u8, 215u8, 11u8, 27u8,
-                            26u8, 231u8, 59u8, 249u8, 78u8, 172u8, 66u8, 81u8, 17u8, 99u8, 221u8,
-                            38u8, 253u8, 62u8, 54u8, 104u8, 161u8, 129u8, 92u8, 218u8, 193u8,
-                        ],
-                    )
-                }
-                #[doc = " A lookup for a number of underperformance sessions for a given validator"]
-                pub fn underperformed_validator_session_count(
-                    &self,
-                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "UnderperformedValidatorSessionCount",
-                        vec![::subxt::storage::address::StorageMapKey::new(
-                            _0.borrow(),
-                            ::subxt::storage::address::StorageHasher::Twox64Concat,
-                        )],
-                        [
-                            46u8, 74u8, 46u8, 159u8, 162u8, 118u8, 159u8, 155u8, 233u8, 63u8,
-                            101u8, 201u8, 56u8, 204u8, 126u8, 242u8, 131u8, 5u8, 29u8, 132u8, 43u8,
-                            205u8, 168u8, 157u8, 29u8, 183u8, 127u8, 202u8, 25u8, 245u8, 137u8,
-                            67u8,
-                        ],
-                    )
-                }
-                #[doc = " A lookup for a number of underperformance sessions for a given validator"]
-                pub fn underperformed_validator_session_count_root(
-                    &self,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                    (),
-                    ::subxt::storage::address::Yes,
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "UnderperformedValidatorSessionCount",
-                        Vec::new(),
-                        [
-                            46u8, 74u8, 46u8, 159u8, 162u8, 118u8, 159u8, 155u8, 233u8, 63u8,
-                            101u8, 201u8, 56u8, 204u8, 126u8, 242u8, 131u8, 5u8, 29u8, 132u8, 43u8,
-                            205u8, 168u8, 157u8, 29u8, 183u8, 127u8, 202u8, 25u8, 245u8, 137u8,
-                            67u8,
-                        ],
-                    )
-                }
-                #[doc = " Validators to be removed from non reserved list in the next era"]
-                pub fn banned(
-                    &self,
-                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanInfo>,
-                    ::subxt::storage::address::Yes,
-                    (),
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "Banned",
-                        vec![::subxt::storage::address::StorageMapKey::new(
-                            _0.borrow(),
-                            ::subxt::storage::address::StorageHasher::Twox64Concat,
-                        )],
-                        [
-                            202u8, 38u8, 165u8, 35u8, 95u8, 207u8, 116u8, 43u8, 148u8, 73u8, 193u8,
-                            187u8, 1u8, 88u8, 209u8, 13u8, 128u8, 168u8, 121u8, 62u8, 227u8, 172u8,
-                            87u8, 106u8, 15u8, 43u8, 136u8, 240u8, 249u8, 210u8, 25u8, 215u8,
-                        ],
-                    )
-                }
-                #[doc = " Validators to be removed from non reserved list in the next era"]
-                pub fn banned_root(
-                    &self,
-                ) -> ::subxt::storage::address::StaticStorageAddress<
-                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanInfo>,
-                    (),
-                    (),
-                    ::subxt::storage::address::Yes,
-                > {
-                    ::subxt::storage::address::StaticStorageAddress::new(
-                        "Elections",
-                        "Banned",
-                        Vec::new(),
-                        [
-                            202u8, 38u8, 165u8, 35u8, 95u8, 207u8, 116u8, 43u8, 148u8, 73u8, 193u8,
-                            187u8, 1u8, 88u8, 209u8, 13u8, 128u8, 168u8, 121u8, 62u8, 227u8, 172u8,
-                            87u8, 106u8, 15u8, 43u8, 136u8, 240u8, 249u8, 210u8, 25u8, 215u8,
-                        ],
-                    )
-                }
                 #[doc = " Openness of the elections, whether we allow all candidates that bonded enough tokens or"]
                 #[doc = " the validators list is managed by sudo"]
                 pub fn openness(
@@ -6957,40 +6647,6 @@ pub mod api {
             use super::runtime_types;
             pub struct ConstantsApi;
             impl ConstantsApi {
-                #[doc = " Nr of blocks in the session."]
-                pub fn session_period(
-                    &self,
-                ) -> ::subxt::constants::StaticConstantAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                > {
-                    ::subxt::constants::StaticConstantAddress::new(
-                        "Elections",
-                        "SessionPeriod",
-                        [
-                            98u8, 252u8, 116u8, 72u8, 26u8, 180u8, 225u8, 83u8, 200u8, 157u8,
-                            125u8, 151u8, 53u8, 76u8, 168u8, 26u8, 10u8, 9u8, 98u8, 68u8, 9u8,
-                            178u8, 197u8, 113u8, 31u8, 79u8, 200u8, 90u8, 203u8, 100u8, 41u8,
-                            145u8,
-                        ],
-                    )
-                }
-                #[doc = " Maximum acceptable ban reason length."]
-                pub fn maximum_ban_reason_length(
-                    &self,
-                ) -> ::subxt::constants::StaticConstantAddress<
-                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
-                > {
-                    ::subxt::constants::StaticConstantAddress::new(
-                        "Elections",
-                        "MaximumBanReasonLength",
-                        [
-                            98u8, 252u8, 116u8, 72u8, 26u8, 180u8, 225u8, 83u8, 200u8, 157u8,
-                            125u8, 151u8, 53u8, 76u8, 168u8, 26u8, 10u8, 9u8, 98u8, 68u8, 9u8,
-                            178u8, 197u8, 113u8, 31u8, 79u8, 200u8, 90u8, 203u8, 100u8, 41u8,
-                            145u8,
-                        ],
-                    )
-                }
                 #[doc = " The maximum number of winners that can be elected by this `ElectionProvider`"]
                 #[doc = " implementation."]
                 #[doc = ""]
@@ -8188,9 +7844,9 @@ pub mod api {
                         "batch",
                         Batch { calls },
                         [
-                            108u8, 189u8, 67u8, 60u8, 83u8, 95u8, 185u8, 30u8, 130u8, 154u8, 187u8,
-                            99u8, 143u8, 45u8, 132u8, 116u8, 211u8, 119u8, 58u8, 200u8, 4u8, 23u8,
-                            20u8, 193u8, 255u8, 12u8, 3u8, 158u8, 2u8, 45u8, 136u8, 36u8,
+                            122u8, 26u8, 45u8, 204u8, 4u8, 5u8, 48u8, 78u8, 69u8, 239u8, 163u8,
+                            152u8, 58u8, 78u8, 150u8, 163u8, 88u8, 179u8, 30u8, 85u8, 12u8, 194u8,
+                            73u8, 143u8, 107u8, 4u8, 33u8, 120u8, 86u8, 111u8, 200u8, 103u8,
                         ],
                     )
                 }
@@ -8220,10 +7876,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            254u8, 23u8, 88u8, 162u8, 231u8, 17u8, 237u8, 30u8, 157u8, 211u8,
-                            148u8, 103u8, 39u8, 87u8, 225u8, 93u8, 120u8, 34u8, 118u8, 248u8,
-                            128u8, 188u8, 223u8, 110u8, 184u8, 178u8, 138u8, 81u8, 133u8, 75u8,
-                            7u8, 167u8,
+                            185u8, 243u8, 37u8, 220u8, 244u8, 43u8, 118u8, 188u8, 199u8, 172u8,
+                            165u8, 159u8, 222u8, 97u8, 129u8, 96u8, 88u8, 118u8, 250u8, 150u8,
+                            158u8, 137u8, 209u8, 7u8, 95u8, 153u8, 75u8, 84u8, 18u8, 236u8, 214u8,
+                            22u8,
                         ],
                     )
                 }
@@ -8250,9 +7906,9 @@ pub mod api {
                         "batch_all",
                         BatchAll { calls },
                         [
-                            138u8, 66u8, 189u8, 225u8, 195u8, 103u8, 222u8, 116u8, 132u8, 62u8,
-                            15u8, 128u8, 68u8, 12u8, 89u8, 62u8, 57u8, 34u8, 168u8, 22u8, 3u8, 1u8,
-                            209u8, 68u8, 53u8, 167u8, 184u8, 104u8, 233u8, 233u8, 235u8, 57u8,
+                            114u8, 112u8, 31u8, 182u8, 103u8, 131u8, 99u8, 199u8, 93u8, 39u8, 25u8,
+                            201u8, 18u8, 35u8, 22u8, 231u8, 21u8, 215u8, 53u8, 187u8, 181u8, 214u8,
+                            84u8, 157u8, 242u8, 223u8, 172u8, 213u8, 253u8, 231u8, 187u8, 30u8,
                         ],
                     )
                 }
@@ -8279,10 +7935,9 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            109u8, 167u8, 115u8, 62u8, 36u8, 236u8, 215u8, 173u8, 158u8, 114u8,
-                            152u8, 35u8, 233u8, 229u8, 91u8, 220u8, 120u8, 133u8, 134u8, 14u8,
-                            129u8, 195u8, 194u8, 188u8, 238u8, 182u8, 45u8, 124u8, 22u8, 2u8, 36u8,
-                            247u8,
+                            56u8, 1u8, 102u8, 9u8, 162u8, 186u8, 73u8, 154u8, 22u8, 35u8, 197u8,
+                            92u8, 204u8, 232u8, 200u8, 115u8, 131u8, 176u8, 38u8, 103u8, 35u8,
+                            224u8, 253u8, 11u8, 1u8, 0u8, 121u8, 61u8, 213u8, 248u8, 138u8, 63u8,
                         ],
                     )
                 }
@@ -8309,9 +7964,10 @@ pub mod api {
                         "force_batch",
                         ForceBatch { calls },
                         [
-                            93u8, 121u8, 5u8, 140u8, 218u8, 192u8, 35u8, 201u8, 135u8, 37u8, 53u8,
-                            43u8, 241u8, 135u8, 231u8, 121u8, 143u8, 93u8, 15u8, 181u8, 170u8,
-                            19u8, 47u8, 17u8, 64u8, 207u8, 153u8, 211u8, 220u8, 55u8, 70u8, 2u8,
+                            17u8, 192u8, 26u8, 31u8, 219u8, 188u8, 228u8, 147u8, 67u8, 98u8, 206u8,
+                            53u8, 198u8, 142u8, 177u8, 1u8, 193u8, 157u8, 139u8, 121u8, 100u8,
+                            38u8, 144u8, 36u8, 114u8, 178u8, 75u8, 225u8, 71u8, 210u8, 158u8,
+                            140u8,
                         ],
                     )
                 }
@@ -8334,9 +7990,10 @@ pub mod api {
                             weight,
                         },
                         [
-                            62u8, 12u8, 150u8, 112u8, 62u8, 64u8, 240u8, 151u8, 177u8, 30u8, 175u8,
-                            55u8, 59u8, 2u8, 193u8, 42u8, 161u8, 7u8, 164u8, 83u8, 92u8, 233u8,
-                            223u8, 225u8, 54u8, 84u8, 159u8, 156u8, 13u8, 251u8, 196u8, 232u8,
+                            85u8, 59u8, 179u8, 21u8, 183u8, 141u8, 241u8, 22u8, 89u8, 33u8, 154u8,
+                            220u8, 172u8, 152u8, 191u8, 5u8, 123u8, 113u8, 33u8, 204u8, 228u8,
+                            110u8, 203u8, 222u8, 16u8, 76u8, 76u8, 185u8, 17u8, 240u8, 214u8,
+                            154u8,
                         ],
                     )
                 }
@@ -8560,9 +8217,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            44u8, 232u8, 219u8, 245u8, 174u8, 148u8, 89u8, 42u8, 231u8, 7u8, 47u8,
-                            131u8, 203u8, 79u8, 194u8, 27u8, 241u8, 122u8, 110u8, 84u8, 70u8, 63u8,
-                            153u8, 24u8, 250u8, 71u8, 141u8, 132u8, 249u8, 93u8, 68u8, 174u8,
+                            184u8, 115u8, 67u8, 6u8, 254u8, 140u8, 164u8, 39u8, 217u8, 248u8,
+                            145u8, 184u8, 235u8, 30u8, 255u8, 147u8, 51u8, 145u8, 46u8, 199u8,
+                            52u8, 225u8, 220u8, 189u8, 98u8, 146u8, 64u8, 190u8, 33u8, 56u8, 229u8,
+                            5u8,
                         ],
                     )
                 }
@@ -8632,9 +8290,9 @@ pub mod api {
                             max_weight,
                         },
                         [
-                            181u8, 14u8, 83u8, 157u8, 20u8, 12u8, 156u8, 53u8, 15u8, 77u8, 62u8,
-                            232u8, 226u8, 154u8, 187u8, 9u8, 68u8, 200u8, 197u8, 176u8, 227u8,
-                            12u8, 32u8, 180u8, 153u8, 31u8, 236u8, 169u8, 156u8, 131u8, 65u8, 69u8,
+                            152u8, 130u8, 1u8, 238u8, 192u8, 165u8, 26u8, 29u8, 79u8, 155u8, 181u8,
+                            90u8, 193u8, 141u8, 147u8, 237u8, 165u8, 218u8, 222u8, 47u8, 25u8, 8u8,
+                            68u8, 142u8, 67u8, 69u8, 202u8, 53u8, 104u8, 65u8, 182u8, 223u8,
                         ],
                     )
                 }
@@ -9046,10 +8704,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            239u8, 228u8, 136u8, 20u8, 82u8, 135u8, 234u8, 100u8, 176u8, 109u8,
-                            25u8, 209u8, 86u8, 79u8, 108u8, 187u8, 90u8, 171u8, 158u8, 161u8,
-                            143u8, 104u8, 2u8, 219u8, 249u8, 132u8, 147u8, 114u8, 218u8, 229u8,
-                            155u8, 179u8,
+                            146u8, 105u8, 148u8, 230u8, 67u8, 228u8, 181u8, 15u8, 168u8, 130u8,
+                            140u8, 12u8, 152u8, 26u8, 115u8, 215u8, 51u8, 146u8, 184u8, 168u8,
+                            245u8, 215u8, 150u8, 177u8, 206u8, 40u8, 175u8, 20u8, 120u8, 147u8,
+                            131u8, 8u8,
                         ],
                     )
                 }
@@ -9076,10 +8734,9 @@ pub mod api {
                             weight,
                         },
                         [
-                            204u8, 109u8, 231u8, 144u8, 219u8, 116u8, 191u8, 123u8, 199u8, 219u8,
-                            220u8, 47u8, 76u8, 102u8, 122u8, 142u8, 68u8, 156u8, 222u8, 200u8,
-                            193u8, 220u8, 205u8, 85u8, 179u8, 22u8, 1u8, 76u8, 46u8, 246u8, 13u8,
-                            243u8,
+                            205u8, 32u8, 180u8, 200u8, 73u8, 35u8, 119u8, 67u8, 145u8, 8u8, 117u8,
+                            141u8, 105u8, 125u8, 156u8, 67u8, 136u8, 113u8, 124u8, 222u8, 77u8,
+                            56u8, 195u8, 178u8, 2u8, 64u8, 123u8, 84u8, 81u8, 45u8, 95u8, 88u8,
                         ],
                     )
                 }
@@ -9138,9 +8795,10 @@ pub mod api {
                             call: ::std::boxed::Box::new(call),
                         },
                         [
-                            79u8, 135u8, 29u8, 144u8, 82u8, 118u8, 47u8, 40u8, 226u8, 63u8, 151u8,
-                            248u8, 68u8, 62u8, 138u8, 238u8, 6u8, 176u8, 175u8, 12u8, 244u8, 100u8,
-                            176u8, 119u8, 48u8, 45u8, 158u8, 208u8, 78u8, 62u8, 204u8, 130u8,
+                            83u8, 19u8, 61u8, 56u8, 146u8, 27u8, 10u8, 124u8, 135u8, 184u8, 53u8,
+                            205u8, 47u8, 168u8, 207u8, 249u8, 80u8, 15u8, 104u8, 129u8, 19u8,
+                            203u8, 220u8, 223u8, 225u8, 198u8, 124u8, 227u8, 229u8, 151u8, 107u8,
+                            220u8,
                         ],
                     )
                 }
@@ -13077,6 +12735,363 @@ pub mod api {
             }
         }
     }
+    pub mod session_ext {
+        use super::{root_mod, runtime_types};
+        #[doc = "Contains one variant per dispatchable that can be called by an extrinsic."]
+        pub mod calls {
+            use super::{root_mod, runtime_types};
+            type DispatchError = runtime_types::sp_runtime::DispatchError;
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            pub struct SetBanConfig {
+                pub minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
+                pub underperformed_session_count_threshold:
+                    ::core::option::Option<::core::primitive::u32>,
+                pub clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
+                pub ban_period: ::core::option::Option<::core::primitive::u32>,
+            }
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            pub struct BanFromCommittee {
+                pub banned: ::subxt::ext::sp_core::crypto::AccountId32,
+                pub ban_reason: ::std::vec::Vec<::core::primitive::u8>,
+            }
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            pub struct CancelBan {
+                pub banned: ::subxt::ext::sp_core::crypto::AccountId32,
+            }
+            pub struct TransactionApi;
+            impl TransactionApi {
+                #[doc = "Sets ban config, it has an immediate effect"]
+                pub fn set_ban_config(
+                    &self,
+                    minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
+                    underperformed_session_count_threshold: ::core::option::Option<
+                        ::core::primitive::u32,
+                    >,
+                    clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
+                    ban_period: ::core::option::Option<::core::primitive::u32>,
+                ) -> ::subxt::tx::StaticTxPayload<SetBanConfig> {
+                    ::subxt::tx::StaticTxPayload::new(
+                        "SessionExt",
+                        "set_ban_config",
+                        SetBanConfig {
+                            minimal_expected_performance,
+                            underperformed_session_count_threshold,
+                            clean_session_counter_delay,
+                            ban_period,
+                        },
+                        [
+                            228u8, 199u8, 170u8, 155u8, 208u8, 190u8, 211u8, 218u8, 105u8, 213u8,
+                            240u8, 152u8, 92u8, 19u8, 164u8, 28u8, 215u8, 145u8, 47u8, 248u8,
+                            219u8, 75u8, 234u8, 78u8, 29u8, 189u8, 35u8, 106u8, 165u8, 76u8, 27u8,
+                            50u8,
+                        ],
+                    )
+                }
+                #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
+                pub fn ban_from_committee(
+                    &self,
+                    banned: ::subxt::ext::sp_core::crypto::AccountId32,
+                    ban_reason: ::std::vec::Vec<::core::primitive::u8>,
+                ) -> ::subxt::tx::StaticTxPayload<BanFromCommittee> {
+                    ::subxt::tx::StaticTxPayload::new(
+                        "SessionExt",
+                        "ban_from_committee",
+                        BanFromCommittee { banned, ban_reason },
+                        [
+                            60u8, 254u8, 80u8, 201u8, 64u8, 189u8, 255u8, 111u8, 14u8, 9u8, 68u8,
+                            177u8, 196u8, 107u8, 10u8, 177u8, 78u8, 134u8, 98u8, 21u8, 179u8, 9u8,
+                            111u8, 185u8, 155u8, 39u8, 148u8, 88u8, 239u8, 16u8, 24u8, 171u8,
+                        ],
+                    )
+                }
+                #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
+                pub fn cancel_ban(
+                    &self,
+                    banned: ::subxt::ext::sp_core::crypto::AccountId32,
+                ) -> ::subxt::tx::StaticTxPayload<CancelBan> {
+                    ::subxt::tx::StaticTxPayload::new(
+                        "SessionExt",
+                        "cancel_ban",
+                        CancelBan { banned },
+                        [
+                            103u8, 192u8, 40u8, 246u8, 206u8, 52u8, 222u8, 51u8, 39u8, 247u8,
+                            220u8, 175u8, 232u8, 31u8, 168u8, 99u8, 206u8, 45u8, 191u8, 161u8,
+                            107u8, 12u8, 112u8, 54u8, 163u8, 170u8, 221u8, 220u8, 122u8, 177u8,
+                            178u8, 246u8,
+                        ],
+                    )
+                }
+            }
+        }
+        #[doc = "\n\t\t\tThe [event](https://docs.substrate.io/main-docs/build/events-errors/) emitted\n\t\t\tby this pallet.\n\t\t\t"]
+        pub type Event = runtime_types::pallet_session_ext::pallet::Event;
+        pub mod events {
+            use super::runtime_types;
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            #[doc = "Ban thresholds for the next era has changed"]
+            pub struct SetBanConfig(pub runtime_types::primitives::BanConfig);
+            impl ::subxt::events::StaticEvent for SetBanConfig {
+                const PALLET: &'static str = "SessionExt";
+                const EVENT: &'static str = "SetBanConfig";
+            }
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            #[doc = "Validators have been banned from the committee"]
+            pub struct BanValidators(
+                pub  ::std::vec::Vec<(
+                    ::subxt::ext::sp_core::crypto::AccountId32,
+                    runtime_types::primitives::BanInfo,
+                )>,
+            );
+            impl ::subxt::events::StaticEvent for BanValidators {
+                const PALLET: &'static str = "SessionExt";
+                const EVENT: &'static str = "BanValidators";
+            }
+        }
+        pub mod storage {
+            use super::runtime_types;
+            pub struct StorageApi;
+            impl StorageApi {
+                #[doc = " A lookup how many blocks a validator produced."]
+                pub fn session_validator_block_count(
+                    &self,
+                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "SessionValidatorBlockCount",
+                        vec![::subxt::storage::address::StorageMapKey::new(
+                            _0.borrow(),
+                            ::subxt::storage::address::StorageHasher::Twox64Concat,
+                        )],
+                        [
+                            186u8, 91u8, 185u8, 144u8, 216u8, 179u8, 157u8, 132u8, 17u8, 247u8,
+                            241u8, 172u8, 32u8, 7u8, 28u8, 60u8, 188u8, 192u8, 64u8, 29u8, 153u8,
+                            100u8, 130u8, 245u8, 189u8, 251u8, 68u8, 161u8, 202u8, 29u8, 153u8,
+                            131u8,
+                        ],
+                    )
+                }
+                #[doc = " A lookup how many blocks a validator produced."]
+                pub fn session_validator_block_count_root(
+                    &self,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
+                    (),
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "SessionValidatorBlockCount",
+                        Vec::new(),
+                        [
+                            186u8, 91u8, 185u8, 144u8, 216u8, 179u8, 157u8, 132u8, 17u8, 247u8,
+                            241u8, 172u8, 32u8, 7u8, 28u8, 60u8, 188u8, 192u8, 64u8, 29u8, 153u8,
+                            100u8, 130u8, 245u8, 189u8, 251u8, 68u8, 161u8, 202u8, 29u8, 153u8,
+                            131u8,
+                        ],
+                    )
+                }
+                #[doc = " Total possible reward per validator for the current era."]
+                pub fn validator_era_total_reward(
+                    &self,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<
+                        runtime_types::pallet_session_ext::ValidatorTotalRewards<
+                            ::subxt::ext::sp_core::crypto::AccountId32,
+                        >,
+                    >,
+                    ::subxt::storage::address::Yes,
+                    (),
+                    (),
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "ValidatorEraTotalReward",
+                        vec![],
+                        [
+                            111u8, 103u8, 48u8, 14u8, 23u8, 139u8, 162u8, 122u8, 212u8, 85u8, 64u8,
+                            188u8, 36u8, 142u8, 80u8, 224u8, 89u8, 63u8, 104u8, 86u8, 51u8, 111u8,
+                            166u8, 53u8, 189u8, 181u8, 240u8, 250u8, 160u8, 128u8, 179u8, 9u8,
+                        ],
+                    )
+                }
+                #[doc = " Current era config for ban functionality, see [`BanConfig`]"]
+                pub fn ban_config(
+                    &self,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanConfig>,
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                    (),
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "BanConfig",
+                        vec![],
+                        [
+                            115u8, 228u8, 135u8, 32u8, 85u8, 156u8, 44u8, 195u8, 215u8, 11u8, 27u8,
+                            26u8, 231u8, 59u8, 249u8, 78u8, 172u8, 66u8, 81u8, 17u8, 99u8, 221u8,
+                            38u8, 253u8, 62u8, 54u8, 104u8, 161u8, 129u8, 92u8, 218u8, 193u8,
+                        ],
+                    )
+                }
+                #[doc = " A lookup for a number of underperformance sessions for a given validator"]
+                pub fn underperformed_validator_session_count(
+                    &self,
+                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "UnderperformedValidatorSessionCount",
+                        vec![::subxt::storage::address::StorageMapKey::new(
+                            _0.borrow(),
+                            ::subxt::storage::address::StorageHasher::Twox64Concat,
+                        )],
+                        [
+                            46u8, 74u8, 46u8, 159u8, 162u8, 118u8, 159u8, 155u8, 233u8, 63u8,
+                            101u8, 201u8, 56u8, 204u8, 126u8, 242u8, 131u8, 5u8, 29u8, 132u8, 43u8,
+                            205u8, 168u8, 157u8, 29u8, 183u8, 127u8, 202u8, 25u8, 245u8, 137u8,
+                            67u8,
+                        ],
+                    )
+                }
+                #[doc = " A lookup for a number of underperformance sessions for a given validator"]
+                pub fn underperformed_validator_session_count_root(
+                    &self,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
+                    (),
+                    ::subxt::storage::address::Yes,
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "UnderperformedValidatorSessionCount",
+                        Vec::new(),
+                        [
+                            46u8, 74u8, 46u8, 159u8, 162u8, 118u8, 159u8, 155u8, 233u8, 63u8,
+                            101u8, 201u8, 56u8, 204u8, 126u8, 242u8, 131u8, 5u8, 29u8, 132u8, 43u8,
+                            205u8, 168u8, 157u8, 29u8, 183u8, 127u8, 202u8, 25u8, 245u8, 137u8,
+                            67u8,
+                        ],
+                    )
+                }
+                #[doc = " Validators to be removed from non reserved list in the next era"]
+                pub fn banned(
+                    &self,
+                    _0: impl ::std::borrow::Borrow<::subxt::ext::sp_core::crypto::AccountId32>,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanInfo>,
+                    ::subxt::storage::address::Yes,
+                    (),
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "Banned",
+                        vec![::subxt::storage::address::StorageMapKey::new(
+                            _0.borrow(),
+                            ::subxt::storage::address::StorageHasher::Twox64Concat,
+                        )],
+                        [
+                            202u8, 38u8, 165u8, 35u8, 95u8, 207u8, 116u8, 43u8, 148u8, 73u8, 193u8,
+                            187u8, 1u8, 88u8, 209u8, 13u8, 128u8, 168u8, 121u8, 62u8, 227u8, 172u8,
+                            87u8, 106u8, 15u8, 43u8, 136u8, 240u8, 249u8, 210u8, 25u8, 215u8,
+                        ],
+                    )
+                }
+                #[doc = " Validators to be removed from non reserved list in the next era"]
+                pub fn banned_root(
+                    &self,
+                ) -> ::subxt::storage::address::StaticStorageAddress<
+                    ::subxt::metadata::DecodeStaticType<runtime_types::primitives::BanInfo>,
+                    (),
+                    (),
+                    ::subxt::storage::address::Yes,
+                > {
+                    ::subxt::storage::address::StaticStorageAddress::new(
+                        "SessionExt",
+                        "Banned",
+                        Vec::new(),
+                        [
+                            202u8, 38u8, 165u8, 35u8, 95u8, 207u8, 116u8, 43u8, 148u8, 73u8, 193u8,
+                            187u8, 1u8, 88u8, 209u8, 13u8, 128u8, 168u8, 121u8, 62u8, 227u8, 172u8,
+                            87u8, 106u8, 15u8, 43u8, 136u8, 240u8, 249u8, 210u8, 25u8, 215u8,
+                        ],
+                    )
+                }
+            }
+        }
+        pub mod constants {
+            use super::runtime_types;
+            pub struct ConstantsApi;
+            impl ConstantsApi {
+                #[doc = " Nr of blocks in the session."]
+                pub fn session_period(
+                    &self,
+                ) -> ::subxt::constants::StaticConstantAddress<
+                    ::subxt::metadata::DecodeStaticType<::core::primitive::u32>,
+                > {
+                    ::subxt::constants::StaticConstantAddress::new(
+                        "SessionExt",
+                        "SessionPeriod",
+                        [
+                            98u8, 252u8, 116u8, 72u8, 26u8, 180u8, 225u8, 83u8, 200u8, 157u8,
+                            125u8, 151u8, 53u8, 76u8, 168u8, 26u8, 10u8, 9u8, 98u8, 68u8, 9u8,
+                            178u8, 197u8, 113u8, 31u8, 79u8, 200u8, 90u8, 203u8, 100u8, 41u8,
+                            145u8,
+                        ],
+                    )
+                }
+            }
+        }
+    }
     pub mod runtime_types {
         use super::runtime_types;
         pub mod aleph_runtime {
@@ -13149,6 +13164,8 @@ pub mod api {
                 NominationPools(runtime_types::pallet_nomination_pools::pallet::Call),
                 #[codec(index = 20)]
                 Identity(runtime_types::pallet_identity::pallet::Call),
+                #[codec(index = 21)]
+                SessionExt(runtime_types::pallet_session_ext::pallet::Call),
             }
             #[derive(
                 :: subxt :: ext :: codec :: Decode,
@@ -13191,6 +13208,8 @@ pub mod api {
                 NominationPools(runtime_types::pallet_nomination_pools::pallet::Event),
                 #[codec(index = 20)]
                 Identity(runtime_types::pallet_identity::pallet::Event),
+                #[codec(index = 21)]
+                SessionExt(runtime_types::pallet_session_ext::pallet::Event),
             }
             #[derive(
                 :: subxt :: ext :: codec :: Decode,
@@ -14718,26 +14737,6 @@ pub mod api {
                         committee_size:
                             ::core::option::Option<runtime_types::primitives::CommitteeSeats>,
                     },
-                    #[codec(index = 1)]
-                    #[doc = "Sets ban config, it has an immediate effect"]
-                    set_ban_config {
-                        minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
-                        underperformed_session_count_threshold:
-                            ::core::option::Option<::core::primitive::u32>,
-                        clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
-                        ban_period: ::core::option::Option<::core::primitive::u32>,
-                    },
-                    #[codec(index = 2)]
-                    #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
-                    ban_from_committee {
-                        banned: ::subxt::ext::sp_core::crypto::AccountId32,
-                        ban_reason: ::std::vec::Vec<::core::primitive::u8>,
-                    },
-                    #[codec(index = 3)]
-                    #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
-                    cancel_ban {
-                        banned: ::subxt::ext::sp_core::crypto::AccountId32,
-                    },
                     #[codec(index = 4)]
                     #[doc = "Set openness of the elections"]
                     set_elections_openness {
@@ -14762,16 +14761,6 @@ pub mod api {
                     NotEnoughNonReservedValidators,
                     #[codec(index = 3)]
                     NonUniqueListOfValidators,
-                    #[codec(index = 4)]
-                    #[doc = "Raised in any scenario [`BanConfig`] is invalid"]
-                    #[doc = "* `performance_ratio_threshold` must be a number in range [0; 100]"]
-                    #[doc = "* `underperformed_session_count_threshold` must be a positive number,"]
-                    #[doc = "* `clean_session_counter_delay` must be a positive number."]
-                    InvalidBanConfig,
-                    #[codec(index = 5)]
-                    #[doc = "Ban reason is too big, ie given vector of bytes is greater than"]
-                    #[doc = "[`Config::MaximumBanReasonLength`]"]
-                    BanReasonTooBig,
                 }
                 #[derive(
                     :: subxt :: ext :: codec :: Decode,
@@ -14790,30 +14779,8 @@ pub mod api {
                         ::std::vec::Vec<::subxt::ext::sp_core::crypto::AccountId32>,
                         runtime_types::primitives::CommitteeSeats,
                     ),
-                    #[codec(index = 1)]
-                    #[doc = "Ban thresholds for the next era has changed"]
-                    SetBanConfig(runtime_types::primitives::BanConfig),
-                    #[codec(index = 2)]
-                    #[doc = "Validators have been banned from the committee"]
-                    BanValidators(
-                        ::std::vec::Vec<(
-                            ::subxt::ext::sp_core::crypto::AccountId32,
-                            runtime_types::primitives::BanInfo,
-                        )>,
-                    ),
                 }
             }
-            #[derive(
-                :: subxt :: ext :: codec :: Decode,
-                :: subxt :: ext :: codec :: Encode,
-                Clone,
-                Debug,
-                Eq,
-                PartialEq,
-            )]
-            pub struct ValidatorTotalRewards<_0>(
-                pub ::subxt::utils::KeyedVec<_0, ::core::primitive::u32>,
-            );
         }
         pub mod pallet_identity {
             use super::runtime_types;
@@ -16693,6 +16660,97 @@ pub mod api {
                     },
                 }
             }
+        }
+        pub mod pallet_session_ext {
+            use super::runtime_types;
+            pub mod pallet {
+                use super::runtime_types;
+                #[derive(
+                    :: subxt :: ext :: codec :: Decode,
+                    :: subxt :: ext :: codec :: Encode,
+                    Clone,
+                    Debug,
+                    Eq,
+                    PartialEq,
+                )]
+                #[doc = "Contains one variant per dispatchable that can be called by an extrinsic."]
+                pub enum Call {
+                    #[codec(index = 1)]
+                    #[doc = "Sets ban config, it has an immediate effect"]
+                    set_ban_config {
+                        minimal_expected_performance: ::core::option::Option<::core::primitive::u8>,
+                        underperformed_session_count_threshold:
+                            ::core::option::Option<::core::primitive::u32>,
+                        clean_session_counter_delay: ::core::option::Option<::core::primitive::u32>,
+                        ban_period: ::core::option::Option<::core::primitive::u32>,
+                    },
+                    #[codec(index = 2)]
+                    #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
+                    ban_from_committee {
+                        banned: ::subxt::ext::sp_core::crypto::AccountId32,
+                        ban_reason: ::std::vec::Vec<::core::primitive::u8>,
+                    },
+                    #[codec(index = 3)]
+                    #[doc = "Schedule a non-reserved node to be banned out from the committee at the end of the era"]
+                    cancel_ban {
+                        banned: ::subxt::ext::sp_core::crypto::AccountId32,
+                    },
+                }
+                #[derive(
+                    :: subxt :: ext :: codec :: Decode,
+                    :: subxt :: ext :: codec :: Encode,
+                    Clone,
+                    Debug,
+                    Eq,
+                    PartialEq,
+                )]
+                #[doc = "\n\t\t\tCustom [dispatch errors](https://docs.substrate.io/main-docs/build/events-errors/)\n\t\t\tof this pallet.\n\t\t\t"]
+                pub enum Error {
+                    #[codec(index = 0)]
+                    #[doc = "Raised in any scenario [`BanConfig`] is invalid"]
+                    #[doc = "* `performance_ratio_threshold` must be a number in range [0; 100]"]
+                    #[doc = "* `underperformed_session_count_threshold` must be a positive number,"]
+                    #[doc = "* `clean_session_counter_delay` must be a positive number."]
+                    InvalidBanConfig,
+                    #[codec(index = 1)]
+                    #[doc = "Ban reason is too big, ie given vector of bytes is greater than"]
+                    #[doc = "[`Config::MaximumBanReasonLength`]"]
+                    BanReasonTooBig,
+                }
+                #[derive(
+                    :: subxt :: ext :: codec :: Decode,
+                    :: subxt :: ext :: codec :: Encode,
+                    Clone,
+                    Debug,
+                    Eq,
+                    PartialEq,
+                )]
+                #[doc = "\n\t\t\tThe [event](https://docs.substrate.io/main-docs/build/events-errors/) emitted\n\t\t\tby this pallet.\n\t\t\t"]
+                pub enum Event {
+                    #[codec(index = 0)]
+                    #[doc = "Ban thresholds for the next era has changed"]
+                    SetBanConfig(runtime_types::primitives::BanConfig),
+                    #[codec(index = 1)]
+                    #[doc = "Validators have been banned from the committee"]
+                    BanValidators(
+                        ::std::vec::Vec<(
+                            ::subxt::ext::sp_core::crypto::AccountId32,
+                            runtime_types::primitives::BanInfo,
+                        )>,
+                    ),
+                }
+            }
+            #[derive(
+                :: subxt :: ext :: codec :: Decode,
+                :: subxt :: ext :: codec :: Encode,
+                Clone,
+                Debug,
+                Eq,
+                PartialEq,
+            )]
+            pub struct ValidatorTotalRewards<_0>(
+                pub ::subxt::utils::KeyedVec<_0, ::core::primitive::u32>,
+            );
         }
         pub mod pallet_staking {
             use super::runtime_types;
@@ -19556,6 +19614,9 @@ pub mod api {
         pub fn identity(&self) -> identity::constants::ConstantsApi {
             identity::constants::ConstantsApi
         }
+        pub fn session_ext(&self) -> session_ext::constants::ConstantsApi {
+            session_ext::constants::ConstantsApi
+        }
     }
     pub struct StorageApi;
     impl StorageApi {
@@ -19621,6 +19682,9 @@ pub mod api {
         pub fn identity(&self) -> identity::storage::StorageApi {
             identity::storage::StorageApi
         }
+        pub fn session_ext(&self) -> session_ext::storage::StorageApi {
+            session_ext::storage::StorageApi
+        }
     }
     pub struct TransactionApi;
     impl TransactionApi {
@@ -19672,6 +19736,9 @@ pub mod api {
         pub fn identity(&self) -> identity::calls::TransactionApi {
             identity::calls::TransactionApi
         }
+        pub fn session_ext(&self) -> session_ext::calls::TransactionApi {
+            session_ext::calls::TransactionApi
+        }
     }
     #[doc = r" check whether the Client you are using is aligned with the statically generated codegen."]
     pub fn validate_codegen<T: ::subxt::Config, C: ::subxt::client::OfflineClientT<T>>(
@@ -19680,9 +19747,9 @@ pub mod api {
         let runtime_metadata_hash = client.metadata().metadata_hash(&PALLETS);
         if runtime_metadata_hash
             != [
-                191u8, 48u8, 69u8, 156u8, 0u8, 234u8, 203u8, 221u8, 166u8, 21u8, 15u8, 202u8,
-                223u8, 93u8, 68u8, 245u8, 104u8, 158u8, 76u8, 167u8, 70u8, 9u8, 196u8, 75u8, 158u8,
-                131u8, 248u8, 255u8, 120u8, 66u8, 210u8, 100u8,
+                187u8, 18u8, 68u8, 56u8, 81u8, 150u8, 56u8, 194u8, 224u8, 211u8, 222u8, 225u8,
+                234u8, 186u8, 11u8, 241u8, 31u8, 179u8, 65u8, 93u8, 170u8, 76u8, 53u8, 239u8, 40u8,
+                125u8, 58u8, 37u8, 49u8, 194u8, 174u8, 120u8,
             ]
         {
             Err(::subxt::error::MetadataError::IncompatibleMetadata)
