+++ aleph-client/src/utility.rs
@@ -5,7 +5,7 @@ use subxt::{blocks::ExtrinsicEvents, ext::sp_runtime::traits::Hash, Config};
 
 use crate::{
     connections::{AsConnection, TxInfo},
-    pallets::{elections::ElectionsApi, staking::StakingApi},
+    pallets::{session_ext::SessionExtApi, staking::StakingApi},
     AlephConfig, BlockHash,
 };
