+++ bin/runtime/Cargo.toml
@@ -60,6 +60,7 @@ sp-io = { default-features = false, git = "https://github.com/Cardinal-Cryptogra
 primitives = { path = "../../primitives", default-features = false }
 pallet-aleph = { path = "../../pallets/aleph", default-features = false }
 pallet-elections = { path = "../../pallets/elections", default-features = false }
+pallet-session-ext = { path = "../../pallets/session-ext", default-features = false }
 
 [build-dependencies]
 substrate-wasm-builder = { git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "aleph-v0.9.38" }
@@ -112,6 +113,7 @@ std = [
     "pallet-contracts/std",
     "pallet-nomination-pools/std",
     "pallet-nomination-pools-runtime-api/std",
+    "pallet-session-ext/std",
 ]
 short_session = ["primitives/short_session"]
 try-runtime = [
