+++ e2e-tests/src/rewards.rs
@@ -7,6 +7,7 @@ use aleph_client::{
         balances::{BalanceUserApi, BalanceUserBatchExtApi},
         elections::{ElectionsApi, ElectionsSudoApi},
         session::{SessionApi, SessionUserApi},
+        session_ext::SessionExtApi,
         staking::{StakingApi, StakingUserApi},
     },
     primitives::{CommitteeSeats, EraValidators},
@@ -16,8 +17,7 @@ use aleph_client::{
 };
 use anyhow::anyhow;
 use log::{debug, info};
-use pallet_elections::LENIENT_THRESHOLD;
-use primitives::{Balance, BlockHash, EraIndex, SessionIndex, TOKEN};
+use primitives::{Balance, BlockHash, EraIndex, SessionIndex, LENIENT_THRESHOLD, TOKEN};
 use sp_runtime::Perquintill;
 
 use crate::{
@@ -124,7 +124,7 @@ fn check_rewards(
     Ok(())
 }
 
-async fn get_node_performance<S: ElectionsApi>(
+async fn get_node_performance<S: ElectionsApi + SessionExtApi>(
     connection: &S,
     account_id: &AccountId,
     before_end_of_session_block_hash: BlockHash,
@@ -152,7 +152,9 @@ async fn get_node_performance<S: ElectionsApi>(
     lenient_performance
 }
 
-pub async fn check_points<S: ElectionsApi + AlephWaiting + BlocksApi + StakingApi>(
+pub async fn check_points<
+    S: ElectionsApi + SessionExtApi + AlephWaiting + BlocksApi + StakingApi,
+>(
     connection: &S,
     session: SessionIndex,
     era: EraIndex,
