+++ e2e-tests/src/test/rewards.rs
@@ -2,6 +2,7 @@ use aleph_client::{
     pallets::{
         elections::ElectionsApi,
         session::SessionApi,
+        session_ext::SessionExtApi,
         staking::{StakingApi, StakingSudoApi},
     },
     primitives::{CommitteeSeats, EraValidators},
@@ -264,7 +265,7 @@ pub async fn change_stake_and_force_new_era() -> anyhow::Result<()> {
 }
 
 async fn check_points_after_force_new_era<
-    S: SignedConnectionApi + BlocksApi + ElectionsApi + AlephWaiting + StakingApi,
+    S: SignedConnectionApi + BlocksApi + ElectionsApi + SessionExtApi + AlephWaiting + StakingApi,
 >(
     connection: &S,
     start_session: SessionIndex,
