+++ pallets/elections/src/lib.rs
@@ -43,10 +43,8 @@ mod tests;
 mod traits;
 
 use codec::{Decode, Encode};
-use frame_support::{log::info, traits::StorageVersion};
-pub use impls::{compute_validator_scaled_total_rewards, LENIENT_THRESHOLD};
+use frame_support::traits::StorageVersion;
 pub use pallet::*;
-use pallets_support::StorageMigration;
 pub use primitives::EraValidators;
 use scale_info::TypeInfo;
 use sp_std::{
@@ -56,7 +54,7 @@ use sp_std::{
 
 pub type TotalReward = u32;
 
-const STORAGE_VERSION: StorageVersion = StorageVersion::new(3);
+const STORAGE_VERSION: StorageVersion = StorageVersion::new(4);
 
 #[derive(Decode, Encode, TypeInfo)]
 pub struct ValidatorTotalRewards<T>(pub BTreeMap<T, TotalReward>);
@@ -67,55 +65,29 @@ pub mod pallet {
         BoundedSupportsOf, ElectionDataProvider, ElectionProvider, ElectionProviderBase, Support,
         Supports,
     };
-    use frame_support::{log, pallet_prelude::*, traits::Get};
-    use frame_system::{
-        ensure_root,
-        pallet_prelude::{BlockNumberFor, OriginFor},
-    };
-    use pallet_session::SessionManager;
-    use primitives::{
-        BanConfig as BanConfigStruct, BanInfo, BanReason, BlockCount, CommitteeSeats,
-        ElectionOpenness, EraIndex, SessionCount,
-    };
-    use sp_runtime::Perbill;
+    use frame_support::{pallet_prelude::*, traits::Get};
+    use frame_system::{ensure_root, pallet_prelude::OriginFor};
+    use primitives::{BannedValidators, CommitteeSeats, ElectionOpenness};
 
     use super::*;
-    use crate::traits::{
-        EraInfoProvider, SessionInfoProvider, ValidatorExtractor, ValidatorRewardsHandler,
-    };
+    use crate::traits::ValidatorProvider;
 
     #[pallet::config]
     pub trait Config: frame_system::Config {
-        /// Something that provides information about ongoing eras.
-        type EraInfoProvider: EraInfoProvider<AccountId = Self::AccountId>;
         type RuntimeEvent: From<Event<Self>> + IsType<<Self as frame_system::Config>::RuntimeEvent>;
         /// Something that provides data for elections.
         type DataProvider: ElectionDataProvider<
             AccountId = Self::AccountId,
             BlockNumber = Self::BlockNumber,
         >;
-        /// Nr of blocks in the session.
-        #[pallet::constant]
-        type SessionPeriod: Get<u32>;
-        /// Handler for managing new session.
-        type SessionManager: SessionManager<<Self as frame_system::Config>::AccountId>;
-        /// Something that provides information about sessions.
-        type SessionInfoProvider: SessionInfoProvider<Self>;
-        /// Something that handles addition of rewards for validators.
-        type ValidatorRewardsHandler: ValidatorRewardsHandler<Self>;
-        /// Something that removes validators from candidates in elections
-        type ValidatorExtractor: ValidatorExtractor<AccountId = Self::AccountId>;
-
-        /// Maximum acceptable ban reason length.
-        #[pallet::constant]
-        type MaximumBanReasonLength: Get<u32>;
-
+        type ValidatorProvider: ValidatorProvider<AccountId = Self::AccountId>;
         /// The maximum number of winners that can be elected by this `ElectionProvider`
         /// implementation.
         ///
         /// Note: This must always be greater or equal to `T::DataProvider::desired_targets()`.
         #[pallet::constant]
         type MaxWinners: Get<u32>;
+        type BannedValidators: BannedValidators<AccountId = Self::AccountId>;
     }
 
     #[pallet::event]
@@ -123,12 +95,6 @@ pub mod pallet {
     pub enum Event<T: Config> {
         /// Committee for the next era has changed
         ChangeValidators(Vec<T::AccountId>, Vec<T::AccountId>, CommitteeSeats),
-
-        /// Ban thresholds for the next era has changed
-        SetBanConfig(BanConfigStruct),
-
-        /// Validators have been banned from the committee
-        BanValidators(Vec<(T::AccountId, BanInfo)>),
     }
 
     #[pallet::pallet]
@@ -136,36 +102,6 @@ pub mod pallet {
     #[pallet::without_storage_info]
     pub struct Pallet<T>(_);
 
-    #[pallet::hooks]
-    impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
-        fn on_runtime_upgrade() -> frame_support::weights::Weight {
-            let on_chain = <Pallet<T> as GetStorageVersion>::on_chain_storage_version();
-            T::DbWeight::get().reads(1)
-                + match on_chain {
-                    _ if on_chain == STORAGE_VERSION => Weight::zero(),
-                    _ if on_chain == StorageVersion::new(0) => {
-                        migrations::v0_to_v1::Migration::<T, Self>::migrate()
-                            + migrations::v1_to_v2::Migration::<T, Self>::migrate()
-                            + migrations::v2_to_v3::Migration::<T, Self>::migrate()
-                    }
-                    _ if on_chain == StorageVersion::new(1) => {
-                        migrations::v1_to_v2::Migration::<T, Self>::migrate()
-                            + migrations::v2_to_v3::Migration::<T, Self>::migrate()
-                    }
-                    _ if on_chain == StorageVersion::new(2) => {
-                        migrations::v2_to_v3::Migration::<T, Self>::migrate()
-                    }
-                    _ => {
-                        log::warn!(
-                            target: "pallet_elections",
-                            "On chain storage version of pallet elections is {:?} but it should not be bigger than 2",
-                            on_chain
-                        );
-                        Weight::zero()
-                    }
-                }
-        }
-    }
     /// Desirable size of a committee, see [`CommitteeSeats`].
     #[pallet::storage]
     pub type CommitteeSize<T> = StorageValue<_, CommitteeSeats, ValueQuery>;
@@ -193,35 +129,6 @@ pub mod pallet {
     #[pallet::storage]
     pub type NextEraNonReservedValidators<T: Config> =
         StorageValue<_, Vec<T::AccountId>, ValueQuery>;
-
-    /// A lookup how many blocks a validator produced.
-    #[pallet::storage]
-    pub type SessionValidatorBlockCount<T: Config> =
-        StorageMap<_, Twox64Concat, T::AccountId, BlockCount, ValueQuery>;
-
-    /// Total possible reward per validator for the current era.
-    #[pallet::storage]
-    pub type ValidatorEraTotalReward<T: Config> =
-        StorageValue<_, ValidatorTotalRewards<T::AccountId>, OptionQuery>;
-
-    /// Default value for ban config, see [`BanConfig`]
-    #[pallet::type_value]
-    pub fn DefaultBanConfig<T: Config>() -> BanConfigStruct {
-        BanConfigStruct::default()
-    }
-
-    /// Current era config for ban functionality, see [`BanConfig`]
-    #[pallet::storage]
-    pub type BanConfig<T> = StorageValue<_, BanConfigStruct, ValueQuery, DefaultBanConfig<T>>;
-
-    /// A lookup for a number of underperformance sessions for a given validator
-    #[pallet::storage]
-    pub type UnderperformedValidatorSessionCount<T: Config> =
-        StorageMap<_, Twox64Concat, T::AccountId, SessionCount, ValueQuery>;
-
-    /// Validators to be removed from non reserved list in the next era
-    #[pallet::storage]
-    pub type Banned<T: Config> = StorageMap<_, Twox64Concat, T::AccountId, BanInfo>;
 
     /// Default value for elections openness.
     #[pallet::type_value]
@@ -270,86 +177,6 @@ pub mod pallet {
             Ok(())
         }
 
-        /// Sets ban config, it has an immediate effect
-        #[pallet::call_index(1)]
-        #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
-        pub fn set_ban_config(
-            origin: OriginFor<T>,
-            minimal_expected_performance: Option<u8>,
-            underperformed_session_count_threshold: Option<u32>,
-            clean_session_counter_delay: Option<u32>,
-            ban_period: Option<EraIndex>,
-        ) -> DispatchResult {
-            ensure_root(origin)?;
-
-            let mut current_committee_ban_config = BanConfig::<T>::get();
-
-            if let Some(minimal_expected_performance) = minimal_expected_performance {
-                ensure!(
-                    minimal_expected_performance <= 100,
-                    Error::<T>::InvalidBanConfig
-                );
-                current_committee_ban_config.minimal_expected_performance =
-                    Perbill::from_percent(minimal_expected_performance as u32);
-            }
-            if let Some(underperformed_session_count_threshold) =
-                underperformed_session_count_threshold
-            {
-                ensure!(
-                    underperformed_session_count_threshold > 0,
-                    Error::<T>::InvalidBanConfig
-                );
-                current_committee_ban_config.underperformed_session_count_threshold =
-                    underperformed_session_count_threshold;
-            }
-            if let Some(clean_session_counter_delay) = clean_session_counter_delay {
-                ensure!(
-                    clean_session_counter_delay > 0,
-                    Error::<T>::InvalidBanConfig
-                );
-                current_committee_ban_config.clean_session_counter_delay =
-                    clean_session_counter_delay;
-            }
-            if let Some(ban_period) = ban_period {
-                ensure!(ban_period > 0, Error::<T>::InvalidBanConfig);
-                current_committee_ban_config.ban_period = ban_period;
-            }
-
-            BanConfig::<T>::put(current_committee_ban_config.clone());
-            Self::deposit_event(Event::SetBanConfig(current_committee_ban_config));
-
-            Ok(())
-        }
-
-        /// Schedule a non-reserved node to be banned out from the committee at the end of the era
-        #[pallet::call_index(2)]
-        #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
-        pub fn ban_from_committee(
-            origin: OriginFor<T>,
-            banned: T::AccountId,
-            ban_reason: Vec<u8>,
-        ) -> DispatchResult {
-            ensure_root(origin)?;
-            let bounded_description: BoundedVec<_, _> = ban_reason
-                .try_into()
-                .map_err(|_| Error::<T>::BanReasonTooBig)?;
-
-            let reason = BanReason::OtherReason(bounded_description);
-            Self::ban_validator(&banned, reason);
-
-            Ok(())
-        }
-
-        /// Schedule a non-reserved node to be banned out from the committee at the end of the era
-        #[pallet::call_index(3)]
-        #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
-        pub fn cancel_ban(origin: OriginFor<T>, banned: T::AccountId) -> DispatchResult {
-            ensure_root(origin)?;
-            Banned::<T>::remove(banned);
-
-            Ok(())
-        }
-
         /// Set openness of the elections
         #[pallet::call_index(4)]
         #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
@@ -370,7 +197,6 @@ pub mod pallet {
         pub non_reserved_validators: Vec<T::AccountId>,
         pub reserved_validators: Vec<T::AccountId>,
         pub committee_seats: CommitteeSeats,
-        pub committee_ban_config: BanConfigStruct,
     }
 
     #[cfg(feature = "std")]
@@ -380,7 +206,6 @@ pub mod pallet {
                 non_reserved_validators: Vec::new(),
                 reserved_validators: Vec::new(),
                 committee_seats: Default::default(),
-                committee_ban_config: Default::default(),
             }
         }
     }
@@ -396,7 +221,6 @@ pub mod pallet {
                 reserved: self.reserved_validators.clone(),
                 non_reserved: self.non_reserved_validators.clone(),
             });
-            <BanConfig<T>>::put(&self.committee_ban_config.clone());
         }
     }
 
@@ -443,17 +267,6 @@ pub mod pallet {
 
             Ok(())
         }
-
-        fn emit_fresh_bans_event() {
-            let active_era = <T as Config>::EraInfoProvider::active_era().unwrap_or(1);
-            let fresh_bans = Banned::<T>::iter()
-                .filter(|(_acc, info)| info.start == active_era + 1)
-                .collect::<Vec<_>>();
-            if !fresh_bans.is_empty() {
-                info!(target: "pallet_elections", "Fresh bans in era {}: {:?}",active_era, fresh_bans);
-                Self::deposit_event(Event::BanValidators(fresh_bans));
-            }
-        }
     }
 
     #[derive(Debug)]
@@ -471,16 +284,6 @@ pub mod pallet {
         NotEnoughReservedValidators,
         NotEnoughNonReservedValidators,
         NonUniqueListOfValidators,
-
-        /// Raised in any scenario [`BanConfig`] is invalid
-        /// * `performance_ratio_threshold` must be a number in range [0; 100]
-        /// * `underperformed_session_count_threshold` must be a positive number,
-        /// * `clean_session_counter_delay` must be a positive number.
-        InvalidBanConfig,
-
-        /// Ban reason is too big, ie given vector of bytes is greater than
-        /// [`Config::MaximumBanReasonLength`]
-        BanReasonTooBig,
     }
 
     impl<T: Config> ElectionProviderBase for Pallet<T> {
@@ -500,10 +303,6 @@ pub mod pallet {
         /// 1) "`NextEraNonReservedValidators` that are staking and are not banned" in case of Permissioned ElectionOpenness
         /// 2) "All staking and not banned validators" in case of Permissionless ElectionOpenness
         fn elect() -> Result<BoundedSupportsOf<Self>, Self::Error> {
-            Self::emit_fresh_bans_event();
-            let active_era = <T as Config>::EraInfoProvider::active_era().unwrap_or(0);
-            let ban_period = BanConfig::<T>::get().ban_period;
-
             let staking_validators = Self::DataProvider::electable_targets(None)
                 .map_err(Self::Error::DataProvider)?
                 .into_iter()
@@ -512,9 +311,8 @@ pub mod pallet {
                 .into_iter()
                 .filter(|v| staking_validators.contains(v))
                 .collect::<BTreeSet<_>>();
-            let banned_validators = Banned::<T>::iter()
-                .filter(|(_, info)| !Self::ban_expired(info.start, ban_period, active_era + 1))
-                .map(|(v, _)| v)
+            let banned_validators = T::BannedValidators::banned()
+                .into_iter()
                 .collect::<BTreeSet<_>>();
             let old_non_reserved_validators = NextEraNonReservedValidators::<T>::get().into_iter();
